+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/spectraldbsubmit/formats/GnpsJsonGenerator.java
@@ -27,7 +27,7 @@
  * Credit to the Du-Lab development team for the initial commitment to the MGF export module.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit;
+package net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.formats;
 
 import javax.json.Json;
 import javax.json.JsonArray;
