+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/spectraldbsubmit/formats/GnpsValues.java
@@ -28,7 +28,7 @@
  * Credit to the Du-Lab development team for the initial commitment to the MGF export module.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit;
+package net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.formats;
 
 /**
  * Specific values for GNPS entries (fields that only allow these values)
